[fix] change int to long in EvenDigits.java.

+++ src/main/java/codejam/april/EvenDigits.java
@@ -20,14 +20,14 @@ public class EvenDigits {
         for (int i = 0; i < T; i++) {
             // get number of ingredients
             // store each optimal distance
-            int D = in.nextInt();
+            long D = in.nextLong();
 
             System.out.println("Case #" + (i + 1) + ": " + getLeastMove(D));
         }
         in.close();
     }
 
-    public static int getLeastMove(int d) {
+    public static long getLeastMove(long d) {
         String stringForm = String.valueOf(d);
         int size = stringForm.length();
         int idx = 0;
@@ -43,31 +43,24 @@ public class EvenDigits {
         char c = stringForm.charAt(idx);
         int firstOddDigit = c - '0';
         if (firstOddDigit == 9) {
-            int sbLess = 0;
-//            for (int i = 0; i < idx; i++) {
-//                sbLess += (stringForm.charAt(i) - '0') * Math.pow(10, size - i - 1);
-//            }
+            long sbLess = 0;
             sbLess += 8 * Math.pow(10, size - idx - 1);
             for (int i = idx + 1; i < size; i++) {
                 sbLess += 8 * Math.pow(10, size - i - 1);
             }
-            return Integer.valueOf(String.valueOf(d).substring(idx)) - sbLess;
+            return Long.valueOf(String.valueOf(d).substring(idx)) - sbLess;
         } else {
-            int sbLess = 0;
-            int sbGreater = 0;
-//            for (int i = 0; i < idx; i++) {
-//                sbGreater += (stringForm.charAt(i) - '0') * Math.pow(10, size - i - 1);
-//            }
-//            sbLess = sbGreater;
+            long sbLess = 0;
+            long sbGreater = 0;
             sbGreater += (firstOddDigit + 1) * Math.pow(10, size - idx - 1);
             sbLess += (firstOddDigit - 1) * Math.pow(10, size - idx - 1);
             for (int i = idx + 1; i < size; i++) {
                 sbLess += 8 * Math.pow(10, size - i - 1);
 
             }
-            int hasChangedD = Integer.valueOf(String.valueOf(d).substring(idx));
-            int toGreater = sbGreater - hasChangedD;
-            int toLess = hasChangedD - sbLess;
+            long hasChangedD = Long.valueOf(String.valueOf(d).substring(idx));
+            long toGreater = sbGreater - hasChangedD;
+            long toLess = hasChangedD - sbLess;
             return Math.min(toGreater, toLess);
         }
     }
